+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Features
     - Added changelog to About/Help page.
     - Header Tool and Server Info Tool fields can now be copied by clicking them, closes #17.
+    - Most table fields can now be selected & copied, closes #18.
 - Changes
     - Flags, Hi-Value, and Lo-Value for SendTable fields now display in hex.
 - Internal